feat: migrate example to next16
