--- example/app/ScaffoldETHProviderInner.tsx
@@ -1,30 +1,0 @@
-"use client";
-
-import { RainbowKitProvider } from "@rainbow-me/rainbowkit";
-import { WagmiProvider } from "wagmi";
-import { Toaster } from "react-hot-toast";
-import { QueryClientProvider, QueryClient } from "@tanstack/react-query";
-import { wagmiConfig } from "./wagmiConfig";
-
-const queryClient = new QueryClient({
-  defaultOptions: {
-    queries: {
-      refetchOnWindowFocus: false,
-    },
-  },
-});
-
-export const ScaffoldEthAppWithProviders = ({
-  children,
-}: {
-  children: React.ReactNode;
-}) => {
-  return (
-    <WagmiProvider config={wagmiConfig}>
-      <QueryClientProvider client={queryClient}>
-        <RainbowKitProvider>{children}</RainbowKitProvider>
-      </QueryClientProvider>
-      <Toaster />
-    </WagmiProvider>
-  );
-};
