+++ example/eslint.config.mjs
@@ -1,0 +1,9 @@
+import nextCoreWebVitals from "eslint-config-next/core-web-vitals";
+import { defineConfig } from "eslint/config";
+
+export default defineConfig([
+  ...nextCoreWebVitals,
+  {
+    ignores: [".next", "next-env.d.ts"],
+  },
+]);
